refactor: Rename `lib/plugins/aws/package/compile/events/alexaSkill.js`
