+++ lib/plugins/aws/package/compile/events/alexa-skill.js
