+++ lib/plugins/index.js
@@ -42,7 +42,7 @@ module.exports = [
   require('./aws/package/compile/events/rabbitmq.js'),
   require('./aws/package/compile/events/msk/index.js'),
   require('./aws/package/compile/events/alb/index.js'),
-  require('./aws/package/compile/events/alexaSkill.js'),
+  require('./aws/package/compile/events/alexa-skill.js'),
   require('./aws/package/compile/events/alexaSmartHome.js'),
   require('./aws/package/compile/events/iot.js'),
   require('./aws/package/compile/events/iotFleetProvisioning.js'),
